Updated tests and examples broken by the code refactoring
